requirements.txt
Django==2.2.10
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-notifications-hq==1.5.0
jsonfield==2.0.2
mysqlclient==1.4.2.post1
numpy==1.15.0
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
